[engine][feat-704][流计算功能优化]

+++ engine-worker/engine-plugins/flink/k8s-flink110-core/src/main/java/org/apache/flink/optimizer/plantranslate/JobGraphGenerator.java
@@ -307,7 +307,9 @@ public class JobGraphGenerator implements Visitor<PlanNode> {
 				throw new FlinkRuntimeException("Could not compress distributed-cache artifacts.", ioe);
 			} finally {
 				try {
-					Files.deleteIfExists(tmpDir);
+					if (tmpDir != null) {
+						Files.deleteIfExists(tmpDir);
+					}
 				} catch (IOException e) {
 					LOG.error("JobGraphGenerator.addUserArtifactEntries error:", e);
 				}
